requirements.txt
fastapi
uvicorn
python-decouple
python-dotenv
SQLAlchemy
pydantic[email]
pytest
httpx